Nha new update insrert car

+++ src/main/java/com/javaweb/repository/CarBrandRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.javaweb.entity.CarBrandEntity;
 
-public interface CarBrandRepository extends JpaRepository<CarBrandEntity,Integer> {
+public interface CarBrandRepository extends JpaRepository<CarBrandEntity,Long> {
 	public List<CarBrandEntity>  findAll();
 	public List<CarBrandEntity>  findByStatus(String status);
